+++ package.json
@@ -10,7 +10,9 @@
     "test": "jest"
   },
   "dependencies": {
+    "@ouroboros/react-native-picker": "^0.3.1",
     "@react-native-community/clipboard": "^1.5.1",
+    "@react-native-picker/picker": "^2.4.10",
     "@react-navigation/bottom-tabs": "^6.5.7",
     "@react-navigation/native": "^6.1.6",
     "@react-navigation/native-stack": "^6.9.12",
