+++ src/Screens/CommonFlow/AnalyticsDetails/AnalyticsDetailsStyle.js
@@ -1,0 +1,221 @@
+import {StyleSheet} from 'react-native';
+import {SetFontSize} from '../../../Utils/FontSize';
+import {colors} from '../../../Utils/colors';
+
+import {
+  widthPercentageToDP as wp,
+  heightPercentageToDP as hp,
+} from 'react-native-responsive-screen';
+import {ConstantKeys} from '../../../Utils/ConstantKey';
+
+export default StyleSheet.create({
+  mainContainer: {flex: 1},
+  container: {
+    flex: 1,
+    backgroundColor: '#FDFDFD',
+  },
+  headerView: {
+    height: hp(4),
+    width: wp(100),
+    backgroundColor: '#3461A6',
+    justifyContent: 'center',
+    paddingLeft: wp(7),
+    marginTop: hp(4),
+  },
+  headerTxt: {
+    fontSize: SetFontSize.setDimension.textSizeMedium,
+    color: colors.WHITE,
+    fontWeight: '600',
+    fontFamily: ConstantKeys.MUKTA,
+  },
+  headerView2: {
+    flexDirection: 'row',
+    height: hp(4),
+    width: wp(100),
+    backgroundColor: 'white',
+    justifyContent: 'flex-start',
+    paddingLeft: wp(7),
+    marginTop: hp(3),
+    alignItems: 'center',
+    marginBottom: hp(1),
+  },
+  headerTxt2: {
+    fontSize: SetFontSize.setDimension.textSizeLarge,
+    color: colors.BLACK,
+    fontWeight: '600',
+    fontFamily: ConstantKeys.MUKTA,
+    marginLeft: wp(3),
+  },
+  leftFlatlist: {
+    marginBottom: hp(6),
+  },
+  mainView: {
+    flexDirection: 'row',
+    marginHorizontal: wp(1),
+    marginVertical: hp(2),
+  },
+  mainLeftView: {
+    paddingVertical: hp(0.1),
+    height: '100%',
+    width: '32%',
+    alignItems: 'center',
+    paddingHorizontal: wp(1),
+  },
+  mainRightView: {
+    height: '100%',
+    width: '68%',
+    borderColor: colors.BLACK,
+    borderWidth: 1,
+    alignItems: 'center',
+    backgroundColor: colors.WHITE,
+  },
+  leftChildView: {height: '100%', width: '100%'},
+  rightChildView: {height: '100%', width: '100%'},
+  leftButton: {
+    backgroundColor: '#E7FFFF',
+    height: hp(6.5),
+    width: '100%',
+    marginBottom: hp(1.5),
+    alignItems: 'center',
+    justifyContent: 'center',
+    borderWidth: 1,
+    borderColor: '#8BC7C7',
+    borderRadius: 7,
+    paddingHorizontal: wp(1),
+  },
+  selectedButton: {
+    backgroundColor: '#3F51B5', // Change the background color for selected button
+  },
+  buttonTxt: {
+    fontSize: SetFontSize.setDimension.textSizeRegular,
+    color: colors.BLACK,
+    fontWeight: '500',
+    fontFamily: ConstantKeys.MUKTA,
+  },
+  selectedButtonText: {
+    color: '#FFFFFF',
+  },
+  rightTopView: {
+    width: '90%',
+    alignItems: 'center',
+  },
+  rightHeadView: {
+    width: '100%',
+    height: hp(4),
+    alignItems: 'center',
+    justifyContent: 'center',
+    marginTop: hp(1),
+    backgroundColor: '#4051B4',
+  },
+  rightHeadTxt: {
+    fontSize: SetFontSize.setDimension.textSizeMedium,
+    color: colors.WHITE,
+    fontWeight: '600',
+    fontFamily: ConstantKeys.MUKTA,
+  },
+  mainDetailsView: {width: '100%'},
+  detailsView: {
+    height: hp(5),
+    width: '100%',
+    flexDirection: 'row',
+    alignItems: 'center',
+    justifyContent: 'space-between',
+  },
+  leftDetails: {width: '50%'},
+  rightDetails: {
+    flexDirection: 'row',
+    width: '50%',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+  downIcon: {marginLeft: wp(5)},
+  detailsTxt: {
+    fontSize: SetFontSize.setDimension.textSize13,
+    color: colors.BLACK,
+    fontWeight: '500',
+    textAlign: 'center',
+    fontFamily: ConstantKeys.MUKTA,
+  },
+  monthTypeTxt: {
+    fontSize: SetFontSize.setDimension.textSizeMedium,
+    color: colors.BLACK,
+    fontWeight: '500',
+    textAlign: 'center',
+    fontFamily: ConstantKeys.MUKTA,
+  },
+  verticalLine: {
+    backgroundColor: '#D9D9D9',
+    height: 25,
+    width: 2,
+  },
+  horizontalLine: {
+    backgroundColor: '#D9D9D9',
+    height: 2,
+    width: '100%',
+    marginTop: hp(0.5),
+  },
+  rightBottomView: {
+    width: '100%',
+    marginTop: hp(5),
+    paddingBottom: hp(10),
+  },
+  tableHeaderView: {
+    marginHorizontal: '1%',
+    width: '98%',
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+  },
+  tableHeaderTxtView: {
+    height: hp(7),
+    width: '19.5%',
+    backgroundColor: '#776ADE',
+    justifyContent: 'center',
+    alignItems: 'center',
+    paddingHorizontal: wp(1),
+  },
+  tableHeadTxt: {
+    fontSize: SetFontSize.setDimension.textSize11,
+    color: colors.WHITE,
+    fontWeight: '400',
+    textAlign: 'center',
+    fontFamily: ConstantKeys.MUKTA,
+  },
+  flatlistView: {
+    width: '100%',
+    marginBottom: hp(2),
+  },
+  bodyHeaderView: {
+    width: '98%',
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    marginTop: hp(0.4),
+    marginHorizontal: '1%',
+  },
+  bodyHeaderTxtView: {
+    height: hp(7),
+    width: '19.5%',
+    backgroundColor: '#EEEEEE',
+    justifyContent: 'center',
+    alignItems: 'center',
+    paddingHorizontal: wp(0.5),
+  },
+  bodyHeaderTxt: {
+    fontSize: SetFontSize.setDimension.textSize11,
+    color: colors.BLACK,
+    fontWeight: '400',
+    textAlign: 'center',
+    fontFamily: ConstantKeys.MUKTA,
+  },
+  bodyHeaderTxt2: {
+    fontSize: SetFontSize.setDimension.textSize11,
+    color: '#0D8F29',
+    fontWeight: '400',
+    textAlign: 'center',
+    fontFamily: ConstantKeys.MUKTA,
+  },
+  pickerstyle: {
+    flex: 1,
+    height: '50%',
+    color: 'black',
+  },
+});
