--- src/Screens/ExpenseAnalytics/ExpenseAnalyticsStyle.js
@@ -1,200 +1,0 @@
-import {StyleSheet} from 'react-native';
-import {SetFontSize} from '../../Utils/FontSize';
-import {colors} from '../../Utils/colors';
-
-import {
-  widthPercentageToDP as wp,
-  heightPercentageToDP as hp,
-} from 'react-native-responsive-screen';
-import {ConstantKeys} from '../../Utils/ConstantKey';
-
-export default StyleSheet.create({
-  mainContainer: {flex: 1},
-  container: {
-    flex: 1,
-    backgroundColor: '#FDFDFD',
-  },
-  cardComponent: {
-    backgroundColor: colors.WHITE,
-    shadowColor: colors.BLACK,
-    shadowOffset: {height: 1, width: 1},
-    shadowOpacity: 1,
-    elevation: 5,
-    shadowRadius: 1,
-    borderRadius: 10,
-    paddingVertical: hp(1.5),
-    paddingHorizontal: wp(6),
-    marginTop: hp(0.5),
-    marginHorizontal: wp(3),
-    marginBottom: hp(2),
-  },
-  headerView: {
-    height: hp(4),
-    width: wp(100),
-    backgroundColor: '#3461A6',
-    justifyContent: 'center',
-    paddingLeft: wp(7),
-    marginTop: hp(4),
-  },
-  headerTxt: {
-    fontSize: SetFontSize.setDimension.textSizeRegular,
-    color: colors.WHITE,
-    fontWeight: '600',
-  },
-  mainView: {
-    borderWidth: 1,
-    backgroundColor: colors.WHITE,
-    borderColor: '#E8E5E5',
-    borderRadius: 10,
-    marginHorizontal: wp(2),
-    marginTop: hp(2.5),
-    marginBottom: hp(-1),
-    paddingHorizontal: wp(2),
-    paddingTop: hp(1),
-  },
-  calender: {
-    alignSelf: 'flex-end',
-  },
-  roundChart: {
-    top: -40,
-    height: hp(40),
-    width: wp(80),
-    alignSelf: 'center',
-  },
-  barView: {
-    top: -15,
-    flexDirection: 'row',
-    marginLeft: wp(5),
-    paddingVertical: hp(2),
-    marginBottom: hp(-4),
-  },
-  barTxtView: {
-    top: -10,
-    flexDirection: 'column',
-    paddingRight: wp(1),
-  },
-  barLinesView: {
-    position: 'absolute',
-    marginLeft: wp(6),
-  },
-  barTxt: {
-    fontSize: SetFontSize.setDimension.textSizeSmall,
-    color: colors.BLACK,
-    fontWeight: '600',
-    marginTop: hp(1),
-    alignSelf: 'flex-end',
-    fontFamily: ConstantKeys.POPPINS,
-    marginRight: wp(1),
-  },
-  hlines: {
-    top: -50,
-    left: -10,
-    position: 'absolute',
-    height: hp(30),
-    width: wp(90),
-  },
-  vlines: {
-    top: -15,
-    height: hp(22),
-    width: wp(25),
-    position: 'absolute',
-    marginLeft: wp(3),
-  },
-  allTxtView: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    marginHorizontal: wp(5),
-    marginTop: hp(4),
-  },
-  txtView: {
-    flexDirection: 'column',
-    alignSelf: 'flex-start',
-  },
-  txtStyle: {
-    height: hp(4.5),
-    width: wp(24),
-    fontSize: SetFontSize.setDimension.textSizeSmall,
-    color: colors.BLACK,
-    fontWeight: '600',
-    fontFamily: ConstantKeys.INTER,
-  },
-  firstLine: {
-    height: hp(1),
-    width: wp(21),
-    backgroundColor: '#D06426',
-    borderRadius: 20,
-  },
-  secondLine: {
-    height: hp(1),
-    width: wp(21),
-    backgroundColor: '#A93DE8',
-    borderRadius: 20,
-  },
-  thirdLine: {
-    height: hp(1),
-    width: wp(21),
-    backgroundColor: '#8DABE1',
-    borderRadius: 20,
-  },
-  tableHeaderView: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    marginHorizontal: wp(1),
-    marginTop: hp(3),
-  },
-  tableHeaderTxtView: {
-    height: hp(7),
-    width: wp(17),
-    backgroundColor: '#776ADE',
-    justifyContent: 'center',
-    alignItems: 'center',
-    paddingHorizontal: wp(1),
-  },
-  tableHeadTxt: {
-    fontSize: SetFontSize.setDimension.textSize11,
-    color: colors.WHITE,
-    fontWeight: '400',
-    textAlign: 'center',
-    fontFamily: ConstantKeys.MUKTA,
-  },
-  bodyHeaderView: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    marginHorizontal: wp(1),
-    marginTop: hp(0.4),
-  },
-  bodyHeaderTxtView: {
-    height: hp(7),
-    width: wp(17),
-    backgroundColor: '#EEEEEE',
-    justifyContent: 'center',
-    alignItems: 'center',
-    paddingHorizontal: wp(0.5),
-  },
-  bodyHeaderTxt: {
-    fontSize: SetFontSize.setDimension.textSize11,
-    color: colors.BLACK,
-    fontWeight: '400',
-    textAlign: 'center',
-    fontFamily: ConstantKeys.MUKTA,
-  },
-  bodyHeaderTxt2: {
-    fontSize: SetFontSize.setDimension.textSizeSmall,
-    color: '#0D8F29',
-    fontWeight: '400',
-    textAlign: 'center',
-    fontFamily: ConstantKeys.MUKTA,
-  },
-  flatlistView: {marginBottom: hp(10)},
-  hLinesView: {
-    flexDirection: 'row',
-    alignItems: 'center',
-  },
-  hImageLines: {
-    height: 3,
-    width: wp(75),
-    backgroundColor: 'lightgrey',
-    marginLeft: wp(1),
-    marginTop: hp(3),
-  },
-});
